# Bit 574 deserialization speed up