--- a/bittensor/utils/tokenizer_utils.py
+++ b/bittensor/utils/tokenizer_utils.py
@@ -877,8 +877,8 @@ def unravel_topk_token_phrases(compact_topk: torch.Tensor, topk: int, ignore_ind
     topk_tensor_legacy = prepend_tensor_legacy(compact_topk, prob_idx, ignore_index_2)
     topk_tensor = prepend_tensor(compact_topk, prob_idx, ignore_index_2)
 
-    print('legacy shape', topk_tensor_legacy.shape)
-    assert torch.all(torch.eq(topk_tensor_legacy, topk_tensor))
+    print('legacy shape', topk_tensor_legacy.shape, topk_tensor.shape)
+    assert torch.all(torch.eq(topk_tensor_legacy, topk_tensor)), 'not eq to legacy'
 
     max_len = topk_tensor.shape[1]
     topk_tensor -= 2  # remove token offset
@@ -911,29 +911,41 @@ def prepend_tensor(compact_topk, prob_idx, ignore_index):
     split_idx = []
     split_idx_partial = prob_idx[1:][phrase_size!= 2]
     split_size = phrase_size[phrase_size != 2]
-    max_len = max(2, max(split_size).item())
-    
+    split_size = torch.cat((split_size, torch.tensor([len(compact_topk) - prob_idx[-1]]))) # make sure the sizes include the last chunk
     # all the indexs to split 
     for idx, size in zip(split_idx_partial, split_size):
-        split_idx += [idx-size, idx]
+        if len(split_idx) > 0 and idx-size == split_idx[-1]:
+            split_idx += [idx]
+            print(idx, split_idx)
+        else:
+            split_idx += [idx-size, idx]
+            print((idx - size, idx), split_idx)
+
+        
 
     # make sure the last of split_idx == prob_idx, so that we dont miss out the last section cut
     if split_idx[-1] != prob_idx[-1]:
         split_idx += [prob_idx[-1]]
-        split_size = torch.cat((split_size, torch.tensor([len(compact_topk) - prob_idx[-1]])))
 
+    max_len = max(2, max(split_size).item(), (len(compact_topk) - prob_idx[-1]))
     split_topk = torch.tensor_split(compact_topk, split_idx)
+    print('prob_idx', [s.item() for s in prob_idx]) 
+    print('split_idx', [s.item() for s in split_idx]) 
+
+    print('split_topk', [len(s) for s in split_topk]) 
 
     # For each sectio of split_topk, if the length of the section is modular 2, then reshape it
     # Append ignore_index to make sure each phrase has the same length (max_len) 
     phrases_list = []
     i = 0
     for s in split_topk:
-        if i != -1 and len(s) != split_size[i]:
+        if (i != -1 and len(s) != split_size[i]) or (i == -1):
             s_reshape = torch.reshape(s, (-1,2))
+            print(len(s), s_reshape.shape)
             ignore = torch.ones((s_reshape.shape[0], max_len-2)) * ignore_index
             s_reshape = torch.cat((s_reshape, ignore), dim = 1) 
-        else: 
+        else:
+            print(len(s))
             i = i + 1 if (i+1 < len(split_size)) else -1
             ignore = torch.ones(max_len - len(s)) * ignore_index
             s_reshape = torch.cat((s, ignore))